# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Car API",
        "description": "API для управления автомобилями, комплектациями, опциями и ценами",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/configurations": {
            "get": {
                "tags": [
                    "Configurations"
                ],
                "summary": "Получить список всех конфигураций",
                "operationId": "ecf557529a783e14266d0c9e51fb28ea",
                "responses": {
                    "200": {
                        "description": "Список конфигураций",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Comfort"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Configurations"
                ],
                "summary": "Создать новую конфигурацию",
                "operationId": "a686fda99be004f135bc70ea4b237df3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "car_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Luxury"
                                    },
                                    "car_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Конфигурация создана",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 3
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Luxury"
                                        },
                                        "car_id": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/configurations/{id}": {
            "get": {
                "tags": [
                    "Configurations"
                ],
                "summary": "Получить конфигурацию по ID",
                "operationId": "603f20b4445da5c154c2dd1cb913471a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID конфигурации",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Конфигурация найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Comfort"
                                        },
                                        "car_id": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Конфигурация не найдена"
                    }
                }
            },
            "put": {
                "tags": [
                    "Configurations"
                ],
                "summary": "Обновить конфигурацию",
                "operationId": "72f7b27458d70875d30c55a866ee53c1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID конфигурации",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "car_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Premium"
                                    },
                                    "car_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Конфигурация обновлена"
                    },
                    "404": {
                        "description": "Конфигурация не найдена"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Configurations"
                ],
                "summary": "Удалить конфигурацию",
                "operationId": "c548b7fd6072a3f525f104c851e72e0e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID конфигурации",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Конфигурация удалена"
                    },
                    "404": {
                        "description": "Конфигурация не найдена"
                    }
                }
            }
        },
...
    },
...
}